+++ src/components/ui/AppBar.tsx
@@ -1,28 +1,26 @@
 import React from 'react'
-import { Anchor, Avatar, Box, Header, Nav, Text } from 'grommet'
-
-const items = [
-    { label: 'Home', href: '/' },
-    { label: 'Sites', href: 'sites' },
-    { label: 'About', href: 'about' },
-]
+import { Anchor, Avatar, Box, Button, Header, Nav, Text } from 'grommet'
+import { Moon, Sun } from 'grommet-icons'
 
 const gravatarSrc = '//s.gravatar.com/avatar/b7fb138d53ba0f573212ccce38a7c43b?s=80'
 
-const AppBar = (props: object) => {
+interface AppBarProps {
+    dark: boolean
+    setDark: React.Dispatch<React.SetStateAction<boolean>>
+}
+const AppBar: React.FC<AppBarProps> = ({ dark, setDark }) => {
     return (
         <Header
             className="App-header"
             background="brand"
             pad={{ left: 'medium', right: 'small', vertical: 'small' }}
             elevation="medium"
-            {...props}
         >
             <Text size="large">My Sites</Text>
             <Nav direction="row">
-                {items.map((item) => (
-                    <Anchor href={item.href} label={item.label} key={item.label} />
-                ))}
+                <Anchor label="Home" href="/" />
+                <Anchor label="Sites" href="sites" />
+                <Anchor label="About" href="about" />
             </Nav>
             <Box direction="row" align="center" gap="small">
                 <Avatar src={gravatarSrc} />
@@ -30,6 +28,19 @@ const AppBar = (props: object) => {
                     Shimi Sun
                 </Anchor>
             </Box>
+            <Button
+                a11yTitle={dark ? 'Switch to Light Mode' : 'Switch to Dark Mode'}
+                icon={dark ? <Moon /> : <Sun />}
+                onClick={() => setDark(!dark)}
+                tip={{
+                    content: (
+                        <Box pad="small" round="small" background={dark ? 'dark-1' : 'light-3'}>
+                            {dark ? 'Switch to Light Mode' : 'Switch to Dark Mode'}
+                        </Box>
+                    ),
+                    plain: true,
+                }}
+            />
         </Header>
     )
 }
