+++ src/layouts/app/App.tsx
@@ -1,35 +1,12 @@
-import React from 'react'
 import ErrorBoundary from 'components/error-boundary'
-import SiteList from 'components/sitelist'
-import {
-    createBrowserRouter,
-    createRoutesFromElements,
-    Route,
-    RouterProvider,
-} from 'react-router-dom'
-import Home from 'pages/home'
-import About from 'pages/about'
-import NotFound from 'pages/not-found'
-
+import MainLayout from 'layouts/MainLayout'
+import React from 'react'
 // layouts
-import RootLayout from 'layouts/RootLayout'
-
-const router = createBrowserRouter(
-    createRoutesFromElements(
-        <Route path="/" element={<RootLayout />}>
-            <Route index element={<Home />} />
-            <Route path="sites" element={<SiteList />} />
-            <Route path="about" element={<About />} />
-
-            <Route path="*" element={<NotFound />} />
-        </Route>
-    )
-)
 
 function App() {
     return (
         <ErrorBoundary>
-            <RouterProvider router={router} />
+            <MainLayout />
         </ErrorBoundary>
     )
 }
