--- src/pages/home/Home.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export default function Home() {
-    return (
-        <div className="home">
-            <h2>Welcome</h2>
-            <p>This is the home page</p>
-        </div>
-    )
-}
