--- src/pages/sites-page/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SitesPage'
